Refactor: 어드민 장소 리스트 조회 쿼리 개선
